refactor(fill): move `FixtureOutput` class to its own moudule
